first v of working video streamer

+++ RestAPI/VideoStreamAPI/db/services/crud.py
@@ -1,4 +1,5 @@
 from ..db_utils import model_to_dict, from_dict
+
 import logging
 logger : logging.Logger = logging.getLogger("app")
 
@@ -7,44 +8,93 @@ class CrudService:
         self.session_factory = session
         self.model = model
 
-    def GetAll(self):
+    def GetAll(self, args = None):
         with self.session_factory() as session:
-            all_rows = session.query(self.model).all()
-            return [model_to_dict(x, include_relationships=True, session=session) for x in all_rows]
+            try:
+                # rows= []
+                # if args is not None:
+
+                #     pass
+                #     # rows = session.query(self.model.genres.in_(args['genre']))
+                # else:
+                rows = session.query(self.model).all()
+                # else:
+                #     # TODO: test
+                #     # Filtering ops
+                #     if args['genre']:
+                        
+                #     if args['series']:
+                #         rows = session.query(self.model.series.in_(args['series']))
+                #     if args['star']:
+                #         rows = session.query(self.model.stars.in_(args['star']))
+                #     if args['director']:
+                #         rows = session.query(self.model.directors.in_(args['director']))
+                    
+                #     sort_column = getattr(self.model, args['sort_by'])
+                #     if args['order'] == 'desc':
+                #         sort_column = sort_column.desc()
+                    
+                #     query = query.order_by(sort_column)
+            except Exception as e:
+                logger.error(f"Failed to get all: {e}")
+                return []
+
+            return [model_to_dict(x, include_relationships=True, session=session) for x in rows]
 
     def Get(self, id):
         with self.session_factory() as session:
             row =  session.get(self.model,id)
+            if row is None:
+                return None
+            return model_to_dict(row,include_relationships=True, session=session)
+
+    def GetAttr(self, attr, value):
+        with self.session_factory() as session:
+            attr_val = getattr(self.model, attr)
+            row = session.query(self.model).filter(attr_val==value).first()
+            if row is None:
+                return None
             return model_to_dict(row,include_relationships=True, session=session)
 
     def Create(self, data):   
         with self.session_factory() as session:
             try:      
                 entity = self.model()
-                entity = from_dict(entity, model, session)
+                entity = from_dict(
+                    obj=entity,
+                    data=data, 
+                    session=session)
                 session.add(entity)    
                 session.commit()
                 return model_to_dict(entity, include_relationships=True, session=session)
-            except:
+            except Exception as e:
+                logger.error(f"Create model: {e}")
                 session.rollback()
                 return None
 
     def Update(self, data):
         with self.session_factory() as session:
-            try:      
-                entity = session.get(ActorModel,  model["id"])
-                entity = from_dict(entity, model, session) 
-                session.add(entity)  
+            try: 
+                entity = session.get(self.model, data['id'])   
+                entity = from_dict(entity, data, session) 
+                # session.update(entity)  
                 session.commit()
                 return model_to_dict(entity, include_relationships=True, session=session)
-            except:
+            except Exception as e:
+                logger.error(f"Update model: {e}")
                 session.rollback()
                 return None
 
     def Delete(self, id):
         with self.session_factory() as session:
-            target = session.get(ActorModel, id)
-            session.delete(target)
-            return True
-
-        return None   
+            try:
+                target = session.get(self.model, id)
+                if target is None:
+                    return False
+                session.delete(target)
+                session.commit()
+                return True
+            except Exception as e:
+                logger.error(f"Delete model: {e}")
+                session.rollback()
+                return False
